requirements.txt
# Core AI and ML libraries
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langgraph>=0.0.20

# Vector stores and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
chromadb>=0.4.0

# Data processing and retrieval
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
numpy>=1.24.0

# API clients for pharmaceutical data
biopython>=1.81
xmltodict>=0.13.0

# Web framework and UI
streamlit>=1.28.0
gradio>=4.0.0

# Reranking and advanced retrieval
rank-bm25>=0.2.2
transformers>=4.30.0
torch>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
tqdm>=4.65.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For production deployment
gunicorn>=21.0.0
uvicorn>=0.23.0
